+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/grammar/Priorities.java
@@ -7,9 +7,6 @@ import java.util.Set;
 
 import org.metaborg.sdf2table.core.Utilities;
 import org.spoofax.interpreter.terms.IStrategoTerm;
-import org.spoofax.terms.StrategoAppl;
-import org.spoofax.terms.StrategoList;
-import org.spoofax.terms.StrategoString;
 
 public class Priorities{
 	/**
@@ -198,79 +195,5 @@ public class Priorities{
 		}
 		
 		return list;
-	}
-	
-	public static void fromStrategoTerm(IStrategoTerm term, Syntax syntax){
-		if(term instanceof StrategoAppl){
-			StrategoAppl app = (StrategoAppl)term;
-			if(app.getName().equals("Priorities")){
-				StrategoList chains = (StrategoList)app.getSubterm(0);
-				
-				for(IStrategoTerm chain : chains){
-					if(chain instanceof StrategoAppl && ((StrategoAppl)chain).getName().equals("Chain")){
-						StrategoList groups = (StrategoList)chain.getSubterm(0);
-						
-						class Ascendant{
-							public int position = -1;
-							public boolean transitive = true;
-							public List<Priority> priorities = new LinkedList<>();
-						}
-						
-						Ascendant ascendant = null;
-						
-						for(IStrategoTerm group : groups){
-							Ascendant a = new Ascendant();
-							
-							if(group instanceof StrategoAppl && ((StrategoAppl)group).getName().equals("NonTransitive")){
-								a.transitive = false;
-								group = group.getSubterm(0);
-							}
-							if(group instanceof StrategoAppl && ((StrategoAppl)group).getName().equals("WithArguments")){
-								StrategoAppl argv = (StrategoAppl)group.getSubterm(1);
-								if(!argv.getName().equals("Default")){
-									System.err.println("Priorities Section: Malformed term: Unknown priority parameter `"+argv.toString()+"'");
-									continue;
-								}
-								StrategoList positions = (StrategoList)argv.getSubterm(0);
-								if(positions.size() > 1){
-									System.err.println("Priorities Section: Malformed term: Unsuported multi positions priority"); // TODO multi-positions priority ?
-									continue;
-								}
-								a.position = Integer.parseInt(((StrategoString)positions.head()).stringValue());
-								group = group.getSubterm(0);
-							}
-							if(group instanceof StrategoAppl && ((StrategoAppl)group).getName().equals("SimpleGroup")){
-								a.priorities.add(new Priority(SyntaxProduction.fromATerm(group.getSubterm(0), syntax), a.position, a.transitive));
-							}else if(group instanceof StrategoAppl && ((StrategoAppl)group).getName().equals("ProdsGroup")){
-								for(IStrategoTerm tp : (StrategoList)group.getSubterm(0)){
-									a.priorities.add(new Priority(SyntaxProduction.fromATerm(tp, syntax), a.position, a.transitive));
-								}
-							}else{
-								System.err.println("Priorities Section: Malformed term: SimpleGroup expected.");
-							}
-							
-							for(Priority p : a.priorities){
-								if(p != null){
-									if(ascendant != null){
-										for(Priority higher : ascendant.priorities){
-											Priority lower = new Priority(p.production(), higher.position(), higher.isTransitive());
-											higher.production().priorities().addLower(lower);
-											p.production().priorities().addHigher(higher);
-										}
-									}
-								}else{
-									System.err.println("Priorities Section: Malformed term: Undefined production.");
-								}
-							}
-							
-							ascendant = a;
-						}
-					}else{
-						System.err.println("Priorities Section: Malformed term: Chain expected.");
-					}
-				}
-
-			}
-		}
 	}
 }
