+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/grammar/SyntaxProduction.java
@@ -653,103 +653,4 @@ public class SyntaxProduction extends Production{
 				0
 				);
 	}
-	
-	public static SyntaxProduction fromATerm(IStrategoTerm term, Syntax syntax){
-		if(term instanceof IStrategoAppl){
-			StrategoAppl app = (StrategoAppl)term;
-			boolean with_cons = false;
-			if(app.getName().equals("SdfProduction") || (with_cons = true && app.getName().equals("SdfProductionWithCons"))){
-				NonTerminal symbol;
-				String cons = "";
-				List<Symbol> sym_list = new ArrayList<>();
-				StrategoAppl tattrs;
-				
-				if(with_cons){
-					// SdfProductionWithCons(SortCons(<type>), Constructor("<cons>"), ...)
-					symbol = (NonTerminal)Symbol.fromStrategoTerm(app.getSubterm(0).getSubterm(0), syntax.symbols());
-					cons = ((StrategoString)app.getSubterm(0).getSubterm(1).getSubterm(0)).stringValue();
-				}else{
-					symbol = (NonTerminal)Symbol.fromStrategoTerm(app.getSubterm(0), syntax.symbols());
-				}
-				
-				// Read right hand side of the equation: Rhs([<symbols>])
-				StrategoList rhs = (StrategoList)app.getSubterm(1).getSubterm(0);
-				for(IStrategoTerm t : rhs){
-					sym_list.add(Symbol.fromStrategoTerm(t, syntax.symbols()));
-				}
-				
-				// Read attributes
-				tattrs = (StrategoAppl)app.getSubterm(2);
-				Set<Attribute> attrs = EnumSet.noneOf(Attribute.class);
-				switch(tattrs.getName()){
-				case "NoAttrs":
-					break;
-				case "Attrs": // TODO handle all production attributes
-					StrategoList talist = (StrategoList)tattrs.getSubterm(0);
-					for(IStrategoTerm ta : talist){
-						StrategoAppl a = (StrategoAppl)ta;
-						switch(a.getName()){
-						case "Assoc":
-							StrategoAppl assoc = (StrategoAppl)a.getSubterm(0);
-							switch(assoc.getName()){
-							case "Left":
-								attrs.add(Attribute.ASSOC_LEFT);
-								break;
-							case "Right":
-								attrs.add(Attribute.ASSOC_RIGHT);
-								break;
-							case "NonAssoc":
-								attrs.add(Attribute.NON_ASSOC);
-								break;
-							default:
-								System.err.println("Unknown associativity: `"+assoc.getName()+"'.");
-								break;
-							}
-							break;
-						case "Bracket":
-							attrs.add(Attribute.BRACKET);
-							break;
-						case "Reject":
-							attrs.add(Attribute.REJECT);
-							break;
-						case "Prefer":
-							attrs.add(Attribute.PREFER);
-							break;
-						case "Avoid":
-							attrs.add(Attribute.AVOID);
-							break;
-						case "CaseInsensitive":
-							attrs.add(Attribute.CASE_INSENSITIVE);
-							break;
-						case "Term":
-							ta = ta.getSubterm(0);
-							if(ta instanceof StrategoAppl && ((StrategoAppl)ta).getName().equals("Default")){
-								ta = ta.getSubterm(0);
-								if(ta instanceof StrategoAppl && ((StrategoAppl)ta).getName().equals("Fun")){
-									ta = ta.getSubterm(0);
-									if(ta instanceof StrategoAppl && ((StrategoAppl)ta).getName().equals("Unquoted")){
-										StrategoString str = (StrategoString)ta.getSubterm(0);
-										if(str.stringValue().equals("longest-match")){
-											attrs.add(Attribute.LONGEST_MATCH);
-											break;
-										}
-									}
-								}
-							}
-						default:
-							System.err.println("Unknown attribute type: `"+a.getName()+"'.");
-							break;
-						}
-					}
-					break;
-				default:
-					System.err.println("Malformed Stratego term : Attributes expected.");
-					return null;
-				}
-				
-				return syntax.uniqueProduction(new SyntaxProduction(symbol, cons, sym_list, attrs));
-			}
-		}
-		return null;
-	}
 }
