+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/grammar/UnloadedModule.java
@@ -1,9 +1,10 @@
 package org.metaborg.sdf2table.grammar;
 
+import org.metaborg.sdf2table.io.Importer;
 import org.spoofax.terms.StrategoAppl;
 
 public class UnloadedModule extends Module{
-	StrategoAppl _syntax_term;
+	StrategoAppl _kernel_term;
 	StrategoAppl _restrictions_term;
 	StrategoAppl _priorities_term;
 	
@@ -12,7 +13,7 @@ public class UnloadedModule extends Module{
 	}
 	
 	public void setSyntaxTerm(StrategoAppl term){
-		_syntax_term = term;
+		_kernel_term = term;
 	}
 	
 	public void setPrioritiesTerm(StrategoAppl term){
@@ -23,15 +24,15 @@ public class UnloadedModule extends Module{
 		_restrictions_term = term;
 	}
 
-	public void loadSyntax(){
-		Syntax.fromStrategoTerm(_syntax_term, syntax());
+	public void loadKernel(){
+		Importer.importKernel(syntax(), _kernel_term);
 	}
 	
 	public void loadRestrictions(){
-		Restrictions.fromStrategoTerm(_restrictions_term, syntax(), symbols());
+		Importer.importRestrictions(syntax(), symbols(), _restrictions_term);
 	}
 	
 	public void loadPriorities(){
-		Priorities.fromStrategoTerm(_priorities_term, syntax());
+		Importer.importPriorities(syntax(), _priorities_term);
 	}
 }
