+++ sdf2table.core/src/main/java/org/metaborg/sdf2table/io/Parenthesizer.java
@@ -1,4 +1,4 @@
-package org.metaborg.sdf2table.parenthesizer;
+package org.metaborg.sdf2table.io;
 
 import java.util.ArrayList;
 import java.util.HashSet;
